- Add copyright information - Move files to identity/openid
